Struct yemantic_ui::collections::button::ButtonProps

pub struct ButtonProps {
    pub root: String,
    pub active: bool,
    pub animated: Either<bool, ButtonAnimationType>,
    pub attached: Either<bool, ButtonAttachedPosition>,
    pub basic: bool,
    pub children: Children,
    pub circular: bool,
    pub class_name: Option<String>,
    pub color: Option<Colors>,
    pub compact: bool,
    pub content: Option<String>,
    pub disabled: bool,
    pub floated: Option<Float>,
    pub fluid: bool,
    pub icon: ChildrenWithProps<Icon>,
    pub inverted: bool,
    pub label: ChildrenWithProps<Label>,
    pub label_position: ButtonLabelPosition,
    pub loading: bool,
    pub negative: bool,
    pub onclick: Callback<MouseEvent>,
    pub positive: bool,
    pub primary: bool,
    pub role: Option<String>,
    pub secondary: bool,
    pub size: Option<Sizes>,
    pub tab_index: Option<isize>,
    pub toggle: bool,
}

Fields
root: String

An html element type to render as root element.

active: bool

A button can show it is currently the active user selection.

animated: Either<bool, ButtonAnimationType>

A button can animate to show hidden content.

attached: Either<bool, ButtonAttachedPosition>

A button can be attached to other content.

basic: bool

A basic button is less pronounced.

children: Children

Primary content.

circular: bool

A button can be circular.

class_name: Option<String>

Additional classes.

color: Option<Colors>

A button can have different colors

compact: bool

A button can reduce its padding to fit into tighter spaces.

content: Option<String>

Shorthand for primary content.

disabled: bool

A button can show it is currently unable to be interacted with.

floated: Option<Float>

A button can be aligned to the left or right of its container.

fluid: bool

A button can take the width of its container.

icon: ChildrenWithProps<Icon>

Add an Icon by pass an .

inverted: bool

A button can be formatted to appear on dark backgrounds.

label: ChildrenWithProps<Label>

Add a Label by pass a .

label_position: ButtonLabelPosition

A labeled button can format a Label or Icon to appear on the left or right.

loading: bool

A button can show a loading indicator.

negative: bool

A button can hint towards a negative consequence.

onclick: Callback<MouseEvent>positive: bool

A button can hint towards a positive consequence.

primary: bool

A button can be formatted to show different levels of emphasis.

role: Option<String>

The role of the HTML element.

secondary: bool

A button can be formatted to show different levels of emphasis.

size: Option<Sizes>

A button can have different sizes.

tab_index: Option<isize>

A button can receive focus.

toggle: bool

A button can be formatted to toggle on and off.